adding CheckLogin and isLogged function.
before adding the function a user was added to users table to insure first log-in is successful.

the password encoded manually with SQL script with this line of code
-- SET password = MD5(password);

after first log-in try, the page sends the user back to the login_page with the correct name and password.

after implementing console.log of the line
Query = `SELECT * FROM users WHERE name = '${name}' AND password = '${enc_pass}'`;

and after comparing both the database's users table's password and the query output notice that the password in the database wasn't encoded with "A" from the start like in this line  md5("A"+password); in CheckLogin function.

 and thus could not establish a proper connection